feat(config): add opt-in [server] table with per-runtime enabled flags
