+++ src/config/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod api;
 pub(crate) mod auth;
 pub(crate) mod config;
+pub(crate) mod server;
